+++ kazoo/protocol/connection.py
@@ -165,7 +165,12 @@ class ConnectionHandler(object):
         with socket_error_handling():
             while remaining > 0:
                 s = self.handler.select([self._socket], [], [], timeout)[0]
-                chunk = s[0].recv(remaining)
+                try:
+                    chunk = s[0].recv(remaining)
+                except IndexError:  # pragma: nocover
+                    # If the read list is empty, we got a timeout. We don't
+                    # have to check wlist and xlist as we don't set any
+                    raise self.handler.timeout_exception("socket time-out")
                 if chunk == b'':
                     raise ConnectionDropped('socket connection broken')
                 msgparts.append(chunk)
@@ -229,7 +234,12 @@ class ConnectionHandler(object):
             while sent < msg_length:
                 s = self.handler.select([], [self._socket], [], timeout)[1]
                 msg_slice = buffer(msg, sent)
-                bytes_sent = s[0].send(msg_slice)
+                try:
+                    bytes_sent = s[0].send(msg_slice)
+                except IndexError:  # pragma: nocover
+                    # If the write list is empty, we got a timeout. We don't
+                    # have to check rlist and xlist as we don't set any
+                    raise self.handler.timeout_exception("socket time-out")
                 if not bytes_sent:
                     raise ConnectionDropped('socket connection broken')
                 sent += bytes_sent