+++ kazoo/tests/test_connection.py
@@ -8,10 +8,12 @@ from nose import SkipTest
 from nose.tools import eq_
 from nose.tools import raises
 
+from kazoo.exceptions import ConnectionLoss
 from kazoo.protocol.serialization import (
     write_string,
     int_struct
 )
+from kazoo.protocol.states import KazooState
 from kazoo.testing import KazooTestCase
 
 
@@ -40,7 +42,6 @@ class TestConnectionHandler(KazooTestCase):
         testit()
 
     def test_with_bad_sessionid(self):
-        from kazoo.protocol.states import KazooState
         ev = threading.Event()
 
         def expired(state):
@@ -56,18 +57,30 @@ class TestConnectionHandler(KazooTestCase):
         client.stop()
 
     def test_connection_dropped(self):
+        # stop default client from interfering
+        self.client.stop()
         client = self._get_client(randomize_hosts=False)
         client.start()
+        ev = threading.Event()
+
+        def back(state):
+            if state == KazooState.CONNECTED:
+                ev.set()
+
         # make sure we are connected to cluster node 0
         eq_(self.cluster[0].server_info.client_port,
             client._connection._socket.getpeername()[1])
-
         # create a node with a large value and stop the ZK node
         path = "/" + uuid.uuid4().hex
-        result = client.create_async(path, b'a' * 1000 * 1024)
+        client.create(path)
+        client.add_listener(back)
+        result = client.set_async(path, b'a' * 1000 * 1024)
         self.cluster[0].stop()
-        node_path = result.get()
-        self.assertTrue(node_path.endswith(path))
+        self.assertRaises(ConnectionLoss, result.get)
+        # we have a working connection to a new node
+        ev.wait(30)
+        eq_(ev.is_set(), True)
+        client.delete(path)
 
 
 class TestReadOnlyMode(KazooTestCase):