+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-0.9 (unreleased)
+0.9 (2013-01-07)
 ----------------
 
 API Changes